// src/types/models/ability.ts
export enum AbilityType {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
}

export enum AbilityName {
  USER_READ = 'USER:READ',
  USER_UPDATE = 'USER:UPDATE',
  USER_TYPE_UPDATE = 'USER_TYPE:UPDATE',
  ROLE_CREATE = 'ROLE:CREATE',
  ROLE_READ = 'ROLE:READ',
  ROLE_UPDATE = 'ROLE:UPDATE',
  ROLE_DELETE = 'ROLE:DELETE',
  CATEGORY_CREATE = 'CATEGORY:CREATE',
  CATEGORY_READ = 'CATEGORY:READ',
  CATEGORY_UPDATE = 'CATEGORY:UPDATE',
  CATEGORY_DELETE = 'CATEGORY:DELETE',
}